+++ HISTORY.rst
@@ -13,6 +13,7 @@
 - Update Docker installer, sample docker-compose file and documentation to
   add a password to the Redis container. GitLab issue #712. Thanks to
   Matthew Thode (@prometheanfire) for the report.
+- Use a fake config file during tests.
 
 3.3.6 (2019-12-19)
 ==================